Refractoring, optimization, and other changes
