+++ src/app/(dashboard)/settings/page.tsx
@@ -1,4 +1,4 @@
-import { SettingsClient } from "./settings-client";
+import { SettingsClient } from "~/app/(dashboard)/settings/settings-client";
 
 export default function SettingsPage() {
   return <SettingsClient />;
