+++ src/app/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import { DashboardClient } from "./client-page";
+import { DashboardClient } from "~/app/dashboard/client-page";
 
 export default function DashboardPage() {
   return <DashboardClient />;
