+++ src/app/prefetch/page.tsx
@@ -1,5 +1,5 @@
 import { serverApi } from "~/trpc/server";
-import PrefetchClientPage from "./client-page";
+import PrefetchClientPage from "~/app/prefetch/client-page";
 
 export default async function Page() {
   const data = await serverApi.example.hello2();
