+++ src/app/template-page/page.tsx
@@ -1,5 +1,5 @@
-import { TemplatePageClient } from "./client-page";
-import { AuthPageWrapper } from "./auth-wrapper";
+import { TemplatePageClient } from "~/app/template-page/client-page";
+import { AuthPageWrapper } from "~/app/template-page/auth-wrapper";
 
 export default function TemplatePage() {
   return (
