+++ src/components/ui/billing-toggle.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from "./button";
+import { Button } from "~/components/ui/button";
 import { cn } from "~/lib/utils";
 
 export type BillingInterval = 'monthly' | 'yearly';
